requirements.txt
appdirs==1.4.3
attrs==19.3.0
black
click
flake8
Jinja2
-e git+https://github.com/colmjude/colmjude-frontend.git#egg=colmjude-frontend
isort
Markdown==3.2.1
markdowntweetembed
MarkupSafe==1.1.1
paramiko
pydantic
python-dotenv==0.13.0
python-frontmatter
PyYAML
regex==2020.4.4
six==1.14.0
sqlite-utils
toml==0.10.0
typed-ast==1.4.1
tweepy
